+++ R/hidden.R
@@ -1,13 +1,10 @@
 #' Initialize a Shiny tag as hidden
 #'
 #' Create a Shiny tag that is invisible when the Shiny app starts. The tag can
-#' be made visible later with \code{shinyjs::toggle} or \code{shinyjs::show}.
+#' be made visible later with [toggle()] or [show()].
 #'
 #' @param ... Shiny tag (or tagList or list of of tags) to make invisible
-#' @seealso \code{\link[shinyjs]{useShinyjs}},
-#' \code{\link[shinyjs]{toggle}},
-#' \code{\link[shinyjs]{show}},
-#' \code{\link[shinyjs]{hide}}
+#' @seealso [useShinyjs()], [toggle()], [show()], [hide()]
 #' @note \code{shinyjs} must be initialized with a call to \code{useShinyjs()}
 #' in the app's ui.
 #' @return The tag (or tags) that was given as an argument in a hidden state.
@@ -36,7 +33,7 @@
 #' hidden(list(span(id = "a"), div(id = "b")))
 #' @export
 hidden <- function(...) {
-  tags <- list(...)
+  tags <- rlang::list2(...)
 
   # recursively add the hidden class to all tags
   if (length(tags) == 1 && inherits(tags[[1]], "shiny.tag")) {