+++ inst/srcjs/shinyjs-default-funcs.js
@@ -1,4 +1,4 @@
-// shinyjs 2.0.0 by Dean Attali
+// shinyjs by Dean Attali
 // Perform common JavaScript operations in Shiny apps using plain R code
 
 shinyjs = function() {
@@ -232,10 +232,14 @@ shinyjs = function() {
       else if ($el.hasClass("shiny-colour-input")) {
         $el = $(_getContainer($el)[0]);
       }
+      // for file inputs, we want to disable the button as well
+      else if ($el.attr('type') === "file") {
+        $el = $el.closest(".shiny-input-container");
+      }
 
       // enable/disable the container as well as all individual inputs inside
       // (this is needed for grouped inputs such as radio and checkbox groups)
-      var toadd = $el.find("input, button, textarea, select");
+      var toadd = $el.find("input, button, textarea, select, a[download], span.btn.btn-file");
       $el = $($el.toArray().concat(toadd.toArray()));
       $el.attr('disabled', (method == "disable"));
       $el.prop('disabled', (method == "disable"));
@@ -321,7 +325,13 @@ shinyjs = function() {
         // sliderInput
         else if (input.children(".js-range-slider").length > 0) {
           input = input.children(".js-range-slider");
-          inputType = "Slider";
+          if (input.attr('data-data-type') == 'date') {
+            inputType = "SliderDate";
+          } else if (input.attr('data-data-type') == 'datetime') {
+            inputType = "SliderDateTime";
+          } else {
+            inputType = "Slider";
+          }
           if (typeof input.attr('data-from') !== "undefined") {
             inputValue = input.attr('data-from');
           }
@@ -462,7 +472,7 @@ shinyjs = function() {
         var eventSimple = {};
         $.each(props, function(idx, prop) {
           if (prop in event) {
-            eventSimple[[prop]] = event[[prop]];
+            eventSimple[prop] = event[prop];
           }
         });
         eventSimple.shinyjsRandom = Math.random();
@@ -737,6 +747,53 @@ shinyjs = function() {
       }
     },
 
+    removeEvent : function (params) {
+      var defaultParams = {
+        event        : null,
+        id           : null
+      }
+      params = shinyjs.getParams(params, defaultParams);
+
+      var el = _jqid(params.id);
+
+      var eventInstance = params.event;
+      var eventType = eventInstance.replace(/(.*)-/,"");
+      var removeAll = (eventInstance == eventType);
+
+      // If the element is not yet in the document, remove it from the onevent queue
+      if (el.length == 0) {
+        if (params.id in _oneventData) {
+          var elementData = _oneventData[params.id];
+          if (removeAll) {
+            delete elementData[eventType];
+          } else {
+            if (eventType in elementData) {
+              var eventDatas = elementData[eventType];
+              var newEvents = eventDatas.filter(function(data) {
+                return data.shinyInputId != eventInstance;
+              });
+              if (newEvents.length == 0) {
+                delete elementData[eventType];
+              } else {
+                elementData[eventType] = newEvents;
+              }
+            }
+          }
+        }
+      } else {
+        var attrName = "data-shinyjs-" + eventType;
+        if (el[0].hasAttribute(attrName)) {
+          if (removeAll) {
+            var attrValue = {};
+          } else {
+            var attrValue = JSON.parse(el.attr(attrName));
+            delete attrValue[eventInstance]
+          }
+          el.attr(attrName, JSON.stringify(attrValue));
+        }
+      }
+    },
+
     // the reset function is also complicated because we need R to tell us
     // what input element or form to reset, then the javascript needs to
     // figure out the correct type and initial value for each input and pass
@@ -748,7 +805,11 @@ shinyjs = function() {
       }
       params = shinyjs.getParams(params, defaultParams);
 
-      var el = _jqid(params.id);
+      if (params.id === "") {
+        var el = $("body");
+      } else {
+        var el = _jqid(params.id);
+      }
 
       // find all the resettable input elements
       var resettables;
@@ -808,14 +869,14 @@ shinyjs = function() {
       if ($el === null) return;
       $el[0].click();
     },
-    
+
     // refresh the page
     refresh : function(params) {
       var defaultParams = {
         id : null
       };
       params = shinyjs.getParams(params, defaultParams);
-      
+
       window.location.reload(false);
     }
   };